pyproject.toml
[tool.poetry]
name = "pycilt"
version = "0.1.0"
description = ""
authors = ["Pritha Gupta <prithag@mail.uni-paderborn.de>"]

[tool.poetry.dependencies]
python =  ">=3.8,<3.11"
numpy = "==1.22.4"
scikit-learn = ">=0.24.0,<0.25.0"
scikit-optimize = ">=0.9.0"
h5py = ">=3.1.0"
pandas = ">=1.1.5"
matplotlib = ">=3.3.3"
future = ">=0.18.2"
openpyxl = ">=3.0.0"
tensorflow = ">=2.9.0"
keras = ">=2.9.0"
torch = ">=1.6"
torchvision = ">=0.7.0"
protobuf = "==3.20.1"
psycopg2 = ">=2.9.5"
psycopg2-binary = ">=2.0"
docopt = "==0.6.2"
dill = ">=0.3.6"
infoselect = "==1.0.2"
tqdm = "==4.64.1"
auto-sklearn = ">=0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
